Renaming and moving VarData into its own file
